requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
faiss-cpu==1.7.4
pinecone-client==2.2.4
python-docx==1.1.0
PyPDF2==3.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
requests==2.31.0
aiofiles==23.2.1 